// src/lib/pb/db-types.ts
type BaseCollectionRecord = {
  id: string;
  created: string;
  updated: string;
};

// https://pocketbase.io/docs/collections/#auth-collection
type AuthCollectionRecord = {
  id: string;
  created: string;
  updated: string;
  username: string;
  email: string;
  emailVisibility: boolean;
  verified: boolean;
};
// ...
export type UtilityShopsResponse = {
  shop_number: string;
  supa_tenant?: string;
  utils?: 'elec' | 'water' | 'both' | 'none';
  order?: number;
  supa_id?: string;
  is_vacant?: boolean;
  tenant?: string;
} & BaseCollectionRecord;

export type UtilityShopsCreate = {
  shop_number: string;
  supa_tenant?: string;
  utils?: 'elec' | 'water' | 'both' | 'none';
  order?: number;
  supa_id?: string;
  is_vacant?: boolean;
  tenant?: string;
};
// ...
export type UtilityTenantsResponse = {
  phone?: string;
  avatar?:string; 
} & AuthCollectionRecord;